REF: misplaced Index.__contains__ tests (#33223)
