+++ pandas/tests/indexes/multi/test_indexing.py
@@ -528,6 +528,12 @@ class TestContains:
         assert "element_not_exit" not in idx
         assert "0 day 09:30:00" in idx
 
+    @pytest.mark.slow
+    def test_large_mi_contains(self):
+        # GH#10645
+        result = MultiIndex.from_arrays([range(10 ** 6), range(10 ** 6)])
+        assert not (10 ** 6, 0) in result
+
 
 def test_timestamp_multiindex_indexer():
     # https://github.com/pandas-dev/pandas/issues/26944
